Gateway config security to separate component
