+++ src/app/gateway/states/gateway-configuration/components/basic/gateway-basic-configuration.component.html
@@ -62,108 +62,7 @@
             </mat-form-field>
           </div>
         </div>
-        <div class="tb-form-panel">
-          <div translate class="tb-form-panel-title">security.security</div>
-          <ng-container formGroupName="security">
-            <tb-toggle-select class="toggle-group" formControlName="type">
-              <tb-toggle-option *ngFor="let securityType of securityTypes | keyvalue"
-                                [value]="securityType.key">{{ securityType.value | translate }}
-              </tb-toggle-option>
-            </tb-toggle-select>
-            <mat-form-field appearance="outline"
-                            *ngIf="basicFormGroup.get('thingsboard.security.type').value.toLowerCase().includes('accesstoken')">
-              <mat-label translate>security.access-token</mat-label>
-              <input matInput formControlName="accessToken"/>
-              <mat-error *ngIf="basicFormGroup.get('thingsboard.security.accessToken').hasError('required')">
-                {{ 'security.access-token-required' | translate }}
-              </mat-error>
-              <tb-copy-button
-                matSuffix
-                miniButton="false"
-                *ngIf="basicFormGroup.get('thingsboard.security.accessToken').value"
-                [copyText]="basicFormGroup.get('thingsboard.security.accessToken').value"
-                tooltipText="{{ 'device.copy-access-token' | translate }}"
-                tooltipPosition="above"
-                icon="content_copy">
-              </tb-copy-button>
-              <mat-icon matIconSuffix style="cursor:pointer;"
-                        matTooltip="{{ 'gateway.hints.token' | translate }}">info_outlined
-              </mat-icon>
-            </mat-form-field>
-            <section *ngIf="basicFormGroup.get('thingsboard.security.type').value === 'usernamePassword'">
-              <div class="tb-form-row no-border no-padding tb-standard-fields">
-                <mat-form-field appearance="outline" class="flex">
-                  <mat-label translate>security.clientId</mat-label>
-                  <input matInput formControlName="clientId"/>
-                  <mat-error *ngIf="basicFormGroup.get('thingsboard.security.clientId').hasError('required')">
-                    {{ 'security.clientId-required' | translate }}
-                  </mat-error>
-                  <tb-copy-button
-                    matSuffix
-                    miniButton="false"
-                    *ngIf="basicFormGroup.get('thingsboard.security.clientId').value"
-                    [copyText]="basicFormGroup.get('thingsboard.security.clientId').value"
-                    tooltipText="{{ 'gateway.copy-client-id' | translate }}"
-                    tooltipPosition="above"
-                    icon="content_copy">
-                  </tb-copy-button>
-                  <mat-icon matIconSuffix style="cursor:pointer;"
-                            matTooltip="{{ 'gateway.hints.client-id' | translate }}">info_outlined
-                  </mat-icon>
-                </mat-form-field>
-                <mat-form-field appearance="outline" class="flex">
-                  <mat-label translate>security.username</mat-label>
-                  <input matInput formControlName="username"/>
-                  <mat-error *ngIf="basicFormGroup.get('thingsboard.security.username').hasError('required')">
-                    {{ 'security.username-required' | translate }}
-                  </mat-error>
-                  <tb-copy-button
-                    matSuffix
-                    miniButton="false"
-                    *ngIf="basicFormGroup.get('thingsboard.security.username').value"
-                    [copyText]="basicFormGroup.get('thingsboard.security.username').value"
-                    tooltipText="{{ 'gateway.copy-username' | translate }}"
-                    tooltipPosition="above"
-                    icon="content_copy">
-                  </tb-copy-button>
-                  <mat-icon matIconSuffix style="cursor:pointer;"
-                            matTooltip="{{ 'gateway.hints.username' | translate }}">info_outlined
-                  </mat-icon>
-                </mat-form-field>
-              </div>
-              <mat-form-field appearance="outline" subscriptSizing="dynamic" style="width: 100%"
-                              *ngIf="basicFormGroup.get('thingsboard.security.type').value === 'usernamePassword'">
-                <mat-label translate>gateway.password</mat-label>
-                <input matInput formControlName="password"/>
-                <tb-copy-button
-                  matSuffix
-                  miniButton="false"
-                  *ngIf="basicFormGroup.get('thingsboard.security.password').value"
-                  [copyText]="basicFormGroup.get('thingsboard.security.password').value"
-                  tooltipText="{{ 'gateway.copy-password' | translate }}"
-                  tooltipPosition="above"
-                  icon="content_copy">
-                </tb-copy-button>
-                <mat-icon matIconSuffix style="cursor:pointer;"
-                          matTooltip="{{ 'gateway.hints.password' | translate }}">info_outlined
-                </mat-icon>
-              </mat-form-field>
-            </section>
-            <tb-error style="margin-top: -12px; display: block;"
-                      *ngIf="basicFormGroup.get('thingsboard.security.type').value === 'usernamePassword'"
-                      [error]="basicFormGroup.get('thingsboard.security').hasError('atLeastOne') ?
-          ('device.client-id-or-user-name-necessary' | translate) : ''"></tb-error>
-            <tb-file-input
-              hint="{{ 'gateway.hints.ca-cert' | translate }}"
-              *ngIf="basicFormGroup.get('thingsboard.security.type').value.toLowerCase().includes('tls')"
-              formControlName="caCert"
-              label="{{ 'security.ca-cert' | translate }}"
-              [allowedExtensions]="'pem,cert,key'"
-              [accept]="'.pem, application/pem,.cert, application/cert, .key,application/key'"
-              dropLabel="{{ 'gateway.drop-file' | translate }}">
-            </tb-file-input>
-          </ng-container>
-        </div>
+        <tb-gateway-security-configuration formControlName="security" [device]="device" (initialCredentialsUpdated)="initialCredentialsUpdated.emit($event)"/>
         <tb-report-strategy *ngIf="withReportStrategy" [defaultValue]="ReportStrategyDefaultValue.Gateway" class="tb-form-panel" formControlName="reportStrategy"/>
       </div>
     </ng-template>
